Fixed test and fixed a bug
The test UploadResponseHandlerTest#getUploadSessionOnProgressingUpload mimicked a server response with a 200/OK response code.
In the REST reference (https://learn.microsoft.com/en-us/graph/api/driveitem-createuploadsession?view=graph-rest-1.0#upload-bytes-to-the-upload-session) it is stated that the response code for uploading more bytes to the upload session is 202/Accepted.
The bug in https://github.com/microsoftgraph/msgraph-sdk-java-core/issues/1517 is fixed.
Previously the method UploadResponseHandler#handleResponse handled differently the response in these two cases:
- response code 200/OK
- all other cases

Now the same method handles differently the response in these two cases:
- response code 200/OK or 201/Created
- all other cases

Everything works if the REST reference is corrected and:
- 200/OK is returned when an item is updated, 201/Created is returned when an item is Created
- 202/Accepted is returned when more bytes are requested