**style** Forbid user formatter to format code
Added a setting fo vscode so that if a contributor has a Java style
formatter at user level, it won't be used to format the files of this
project. This could result in a lot of changes due to reformatting
also with few LOCs changed (happened to me). Maybe a common formatter
should be defined and set?