Enable errorMapping with "4XX" and "5XX" pattern
fix: Failed Responses now work with errorMapping with Key "4XX" and "5XX".

It looks like, there should be code here to satisfy the "4XX" and "5XX" keys to match 400-499 and 500-599 StatusCodes. But it does not work yet, so i implemented it quickly. 

Im new here in such big Projects, so i dont know if im at the right address for this issue. I'm open for feedback tho.